refactor: separate concerns between `MimeApps`, `SystemApps`, and `Config`
