+++ handlr-regex/src/common/desktop_entry.rs
@@ -1,6 +1,5 @@
 use crate::{
-    apps::{MimeApps, SystemApps},
-    config::Config,
+    apps_config::AppsConfig,
     error::{Error, ErrorKind, Result},
 };
 use aho_corasick::AhoCorasick;
@@ -49,9 +48,7 @@ impl DesktopEntry {
     /// Execute the command in `exec` in the given mode and with the given arguments
     pub fn exec(
         &self,
-        config: &Config,
-        mime_apps: &mut MimeApps,
-        system_apps: &SystemApps,
+        apps_config: &mut AppsConfig,
         mode: Mode,
         arguments: Vec<String>,
         selector: &str,
@@ -60,29 +57,13 @@ impl DesktopEntry {
         let supports_multiple =
             self.exec.contains("%F") || self.exec.contains("%U");
         if arguments.is_empty() {
-            self.exec_inner(
-                config,
-                mime_apps,
-                system_apps,
-                vec![],
-                selector,
-                enable_selector,
-            )?
+            self.exec_inner(apps_config, vec![], selector, enable_selector)?
         } else if supports_multiple || mode == Mode::Launch {
-            self.exec_inner(
-                config,
-                mime_apps,
-                system_apps,
-                arguments,
-                selector,
-                enable_selector,
-            )?;
+            self.exec_inner(apps_config, arguments, selector, enable_selector)?;
         } else {
             for arg in arguments {
                 self.exec_inner(
-                    config,
-                    mime_apps,
-                    system_apps,
+                    apps_config,
                     vec![arg],
                     selector,
                     enable_selector,
@@ -96,22 +77,14 @@ impl DesktopEntry {
     /// Internal helper function for `exec`
     fn exec_inner(
         &self,
-        config: &Config,
-        mime_apps: &mut MimeApps,
-        system_apps: &SystemApps,
+        apps_config: &mut AppsConfig,
         args: Vec<String>,
         selector: &str,
         enable_selector: bool,
     ) -> Result<()> {
         let mut cmd = {
-            let (cmd, args) = self.get_cmd(
-                config,
-                mime_apps,
-                system_apps,
-                args,
-                selector,
-                enable_selector,
-            )?;
+            let (cmd, args) =
+                self.get_cmd(apps_config, args, selector, enable_selector)?;
             let mut cmd = Command::new(cmd);
             cmd.args(args);
             cmd
@@ -129,9 +102,7 @@ impl DesktopEntry {
     /// Get the `exec` command, formatted with given arguments
     pub fn get_cmd(
         &self,
-        config: &Config,
-        mime_apps: &mut MimeApps,
-        system_apps: &SystemApps,
+        apps_config: &mut AppsConfig,
         args: Vec<String>,
         selector: &str,
         enable_selector: bool,
@@ -176,12 +147,7 @@ impl DesktopEntry {
         // If the entry expects a terminal (emulator), but this process is not running in one, we
         // launch a new one.
         if self.terminal && !std::io::stdout().is_terminal() {
-            let term_cmd = config.terminal(
-                mime_apps,
-                system_apps,
-                selector,
-                enable_selector,
-            )?;
+            let term_cmd = apps_config.terminal(selector, enable_selector)?;
             exec = shlex::split(&term_cmd)
                 .ok_or_else(|| Error::from(ErrorKind::BadCmd(term_cmd)))?
                 .into_iter()
